streamlit/requirements.txt
matplotlib==3.9.2
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
scipy==1.14.1
seaborn==0.13.2
ultralytics==8.2.100
ultralytics-thop==2.0.8
pillow == 10.4.0